Dev/robin/9587 support public and shared assets (#7)
* fix: support events on public and shared assets

depends on upstream change to go-datatrails-simplehash

AB#9587

* feat: events-verify variant of ediag

Also,

* read event data from stdin if a filename is not provided
* provide integration tests that take advantage of the production public
  attestation tenant for edaig and event-verify
* Update the README to include a verification example

AB#9587

* build: bump go-datatrails-simplehash to get direct V3Event hashing

* review: example of reading event data for verification from file

* quality: linter

---------

+++ node_test.go
@@ -32,7 +32,7 @@ func TestNodeCmd(t *testing.T) {
 
 	tenantID := mmrtesting.DefaultGeneratorTenantIdentity
 	testContext, testGenerator, cfg := veracitytesting.NewAzuriteTestContext(t, "TestNodeCmd")
-	veracitytesting.GenerateTenantLog(&testContext, testGenerator, 10, tenantID, true, massifHeight, uint8(LeafTypePlain))
+	veracitytesting.GenerateTenantLog(&testContext, testGenerator, 10, tenantID, true, massifHeight, LeafTypePlain)
 
 	tests := []struct {
 		testArgs []string
